requirements.txt
asgiref==3.11.0
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
Django==6.0
fonttools==4.61.1
joblib==1.5.3
kiwisolver==1.4.9
liac-arff==2.5.0
matplotlib==3.10.8
nltk==3.9.2
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.11.3
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.3
whitenoise==6.11.0
